refactor : create NumberGenerator Interface
