+++ src/main/java/baseball/engine/NumberGenerator.java
@@ -1,20 +1,7 @@
 package baseball.engine;
 
 import baseball.model.Numbers;
-import camp.nextstep.edu.missionutils.Randoms;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
-public class NumberGenerator {
-    private static final int START_RANGE = 1;
-    private static final int END_RANGE = 9;
-
-    public Numbers generate(int size) {
-        return new Numbers(
-                Stream.generate(() -> Randoms.pickNumberInRange(START_RANGE, END_RANGE))
-                        .distinct()
-                        .limit(size)
-                        .collect(Collectors.toList())
-        );
-    }
+public interface NumberGenerator {
+    Numbers generate(int size);
 }
